Implement CRD operations for product

+++ server/src/main/java/com/finalproject/storemanagementproject/controllers/ProductController.java
@@ -1,16 +1,18 @@
 package com.finalproject.storemanagementproject.controllers;
 
 import com.finalproject.storemanagementproject.models.APIResponse;
+import com.finalproject.storemanagementproject.models.OrderProduct;
 import com.finalproject.storemanagementproject.models.Product;
+import com.finalproject.storemanagementproject.services.OrderProductService;
 import com.finalproject.storemanagementproject.services.ProductService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
+import java.time.Instant;
 import java.util.Collections;
 import java.util.List;
-import java.util.Optional;
 
 @RestController
 @RequestMapping("product")
@@ -19,9 +21,16 @@ public class ProductController {
     @Autowired
     private ProductService productService;
 
+    @Autowired
+    OrderProductService orderProductService;
+
     @GetMapping("")
-    public ResponseEntity<APIResponse<Product>> getAllProducts() {
-        List<Product> products = productService.getAllProducts();
+    public ResponseEntity<APIResponse<Product>> getProducts(@RequestParam(required = false) String text) {
+        List<Product> products;
+
+        if (text != null && !text.isEmpty())
+            products = productService.findProductByName(text);
+        else products = productService.getAllProducts();
 
         if (products == null) {
             return ResponseEntity.ok(
@@ -33,11 +42,10 @@ public class ProductController {
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<APIResponse<Optional<Product>>> getProductById(@PathVariable String id) {
+    public ResponseEntity<APIResponse<Product>> getProductById(@PathVariable String id) {
+        Product product = productService.getProductById(id);
 
-        Optional<Product> product = productService.getProductById(id);
-
-        Integer HTTP_CODE = HttpStatus.OK.value();
+        int HTTP_CODE = HttpStatus.OK.value();
         String message = "Success";
 
         if (product == null) {
@@ -50,23 +58,46 @@ public class ProductController {
 
     @PostMapping("/create")
     public ResponseEntity<APIResponse<Product>> createProduct(@RequestBody Product product) {
+        Integer HTTP_CODE = HttpStatus.OK.value();
+        String message = "Create Success";
+
+        if (productService.findByBarCode(product.getBarcode()) != null) {
+            HTTP_CODE = HttpStatus.BAD_REQUEST.value();
+            message = "Barcode is already in use";
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(new APIResponse<>(HTTP_CODE, message, Collections.singletonList(null)));
+        }
+
+        product.setCreatedAt(Instant.now());
+        product.setUpdatedAt(Instant.now());
+
         Product savedProduct = productService.saveProduct(product);
-        return ResponseEntity.ok(new APIResponse<>(HttpStatus.CREATED.value(), "Product has been created successfully",
+        return ResponseEntity.ok(new APIResponse<>(HTTP_CODE, message,
                 Collections.singletonList(savedProduct)));
     }
 
     @PostMapping("update/{id}")
     public ResponseEntity<APIResponse<Product>> updateProduct(@PathVariable String id, @RequestBody Product updatedProduct) {
-        Product product = productService.getProductById(id).orElse(null);
+        Product product = productService.getProductById(id);
 
         Integer HTTP_CODE = HttpStatus.OK.value();
         String message = "Update Success";
+
+        if (productService.findByBarCode(updatedProduct.getBarcode()) != null && !product.getBarcode().equals(updatedProduct.getBarcode())) {
+            HTTP_CODE = HttpStatus.BAD_REQUEST.value();
+            message = "Barcode is already in use";
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(new APIResponse<>(HTTP_CODE, message, Collections.singletonList(null)));
+        }
 
         if (product != null) {
             product.setName(updatedProduct.getName());
             product.setCategory(updatedProduct.getCategory());
             product.setRetailPrice(updatedProduct.getRetailPrice());
-            product.setUpdatedAt(updatedProduct.getUpdatedAt());
+            product.setImportPrice(updatedProduct.getImportPrice());
+            product.setBarcode(updatedProduct.getBarcode());
+            product.setQuantity(updatedProduct.getQuantity());
+            product.setIllustrator(updatedProduct.getIllustrator());
+
+            product.setUpdatedAt(Instant.now());
             productService.saveProduct(product);
             return ResponseEntity.ok(new APIResponse<>(HTTP_CODE, message, Collections.singletonList(product)));
         }
@@ -74,12 +105,25 @@ public class ProductController {
         HTTP_CODE = HttpStatus.NOT_FOUND.value();
         message = "Not found product";
 
-        return ResponseEntity.ok(new APIResponse<>(HTTP_CODE, message, Collections.singletonList(product)));
+        return ResponseEntity.status(HttpStatus.NOT_FOUND).body(new APIResponse<>(HTTP_CODE, message, Collections.singletonList(null)));
     }
 
     @PostMapping("delete/{id}")
-    public ResponseEntity<APIResponse<Void>> deleteProduct(@PathVariable String id) {
-        productService.deleteProduct(id);
-        return ResponseEntity.ok(new APIResponse<>(HttpStatus.OK.value(), "Delete Success", null));
+    public ResponseEntity<APIResponse<Product>> deleteProduct(@PathVariable String id) {
+        try {
+            Product product = productService.getProductById(id);
+            if (product == null)
+                return ResponseEntity.status(HttpStatus.NOT_FOUND).body(new APIResponse<>(HttpStatus.NOT_FOUND.value(), "Not found product", Collections.emptyList()));
+
+            List<OrderProduct> orderProducts = orderProductService.getOrderProductsByPid(id);
+
+            if (orderProducts != null && !orderProducts.isEmpty())
+                return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(new APIResponse<>(HttpStatus.BAD_REQUEST.value(), "Product is in use", Collections.emptyList()));
+
+            productService.deleteProduct(id);
+            return ResponseEntity.ok(new APIResponse<>(HttpStatus.OK.value(), "Delete Success", Collections.emptyList()));
+        } catch (Exception e) {
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(new APIResponse<>(HttpStatus.BAD_REQUEST.value(), "Delete Failed", Collections.emptyList()));
+        }
     }
 }
